+++ packages/react/src/components/video/IosVideoPlayer.tsx
@@ -1,5 +1,4 @@
 import './plyr.css';
-import './styles.css';
 
 import clsx from 'clsx';
 import type { Level } from 'hls.js';