+++ packages/react/src/index.ts
@@ -2,3 +2,4 @@ export * from './hooks';
 export * from './providers';
 export * from './components';
 export * from './types';
+export * from './classes';