+++ tests/specs/freeport-api.spec.ts
@@ -5,7 +5,7 @@ const { nftId } = mockNft;
 const { address: collectionAddress } = mockCollection;
 const { address: walletAddress } = mockSigner;
 
-describe('Freeport API Client', () => {
+describe.skip('Freeport API Client', () => {
   describe('create', () => {
     it('should instantiate a new client', async () => {
       const client = await FreeportApiService.create();
@@ -74,7 +74,7 @@ describe('Freeport API Client', () => {
     });
   });
 
-  describe('getContentDek', () => {
+  describe.skip('getContentDek', () => {
     it.skip('should return a valid DEK if authenticated', async () => {
       const client = await FreeportApiService.create();
       await client.authenticate(mockSigner);
@@ -94,7 +94,7 @@ describe('Freeport API Client', () => {
     });
   });
 
-  describe('getContent', () => {
+  describe.skip('getContent', () => {
     it.skip("should return the NFT's content if authenticated", async () => {
       const client = await FreeportApiService.create();
       await client.authenticate(mockSigner);