+++ packages/client/CHANGELOG.md
@@ -3,27 +3,91 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [1.5.0](https://github.com/Cerebellum-Network/cere-media-sdk-js/compare/v1.4.2-rc.16...v1.5.0) (2024-10-01)
+## [1.6.0-rc.13](https://github.com/Cerebellum-Network/cere-media-sdk-js/compare/v1.6.0-rc.12...v1.6.0-rc.13) (2024-12-19)
 
 **Note:** Version bump only for package @cere/media-sdk-client
+
+## [1.6.0-rc.12](https://github.com/Cerebellum-Network/cere-media-sdk-js/compare/v1.6.0-rc.11...v1.6.0-rc.12) (2024-12-19)
+
+**Note:** Version bump only for package @cere/media-sdk-client
+
+## [1.6.0-rc.11](https://github.com/Cerebellum-Network/cere-media-sdk-js/compare/v1.6.0-rc.10...v1.6.0-rc.11) (2024-12-04)
+
+**Note:** Version bump only for package @cere/media-sdk-client
+
+## [1.6.0-rc.10](https://github.com/Cerebellum-Network/cere-media-sdk-js/compare/v1.6.0-rc.9...v1.6.0-rc.10) (2024-11-28)
+
+**Note:** Version bump only for package @cere/media-sdk-client
+
+## [1.6.0-rc.9](https://github.com/Cerebellum-Network/cere-media-sdk-js/compare/v1.6.0-rc.8...v1.6.0-rc.9) (2024-11-22)
+
+**Note:** Version bump only for package @cere/media-sdk-client
+
+## [1.6.0-rc.8](https://github.com/Cerebellum-Network/cere-media-sdk-js/compare/v1.6.0-rc.7...v1.6.0-rc.8) (2024-11-21)
+
+**Note:** Version bump only for package @cere/media-sdk-client
+
+## [1.6.0-rc.7](https://github.com/Cerebellum-Network/cere-media-sdk-js/compare/v1.6.0-rc.6...v1.6.0-rc.7) (2024-11-21)
+
+**Note:** Version bump only for package @cere/media-sdk-client
+
+## [1.6.0-rc.6](https://github.com/Cerebellum-Network/cere-media-sdk-js/compare/v1.6.0-rc.5...v1.6.0-rc.6) (2024-11-21)
+
+**Note:** Version bump only for package @cere/media-sdk-client
+
+## [1.6.0-rc.5](https://github.com/Cerebellum-Network/cere-media-sdk-js/compare/v1.6.0-rc.4...v1.6.0-rc.5) (2024-11-15)
+
+**Note:** Version bump only for package @cere/media-sdk-client
+
+## [1.6.0-rc.4](https://github.com/Cerebellum-Network/cere-media-sdk-js/compare/v1.6.0-rc.3...v1.6.0-rc.4) (2024-11-14)
+
+**Note:** Version bump only for package @cere/media-sdk-client
+
+## [1.6.0-rc.3](https://github.com/Cerebellum-Network/cere-media-sdk-js/compare/v1.6.0-rc.2...v1.6.0-rc.3) (2024-11-14)
+
+**Note:** Version bump only for package @cere/media-sdk-client
+
+## [1.6.0-rc.2](https://github.com/Cerebellum-Network/cere-media-sdk-js/compare/v1.6.0-rc.1...v1.6.0-rc.2) (2024-11-14)
+
+**Note:** Version bump only for package @cere/media-sdk-client
+
+## [1.6.0-rc.1](https://github.com/Cerebellum-Network/cere-media-sdk-js/compare/v1.6.0-rc.0...v1.6.0-rc.1) (2024-11-14)
+
+**Note:** Version bump only for package @cere/media-sdk-client
+
+## [1.6.0-rc.0](https://github.com/Cerebellum-Network/cere-media-sdk-js/compare/v1.5.2-rc.0...v1.6.0-rc.0) (2024-11-14)
+
+## [1.5.0](https://github.com/Cerebellum-Network/cere-media-sdk-js/compare/v1.4.2-rc.16...v1.5.0) (2024-10-01)
 
 ## [1.4.2-rc.16](https://github.com/Cerebellum-Network/cere-media-sdk-js/compare/v1.4.2-rc.15...v1.4.2-rc.16) (2024-08-29)
 
-**Note:** Version bump only for package @cere/media-sdk-client
-
 ## [1.4.2-rc.15](https://github.com/Cerebellum-Network/cere-media-sdk-js/compare/v1.4.2-rc.14...v1.4.2-rc.15) (2024-08-29)
-
-**Note:** Version bump only for package @cere/media-sdk-client
 
 ## [1.4.2-rc.14](https://github.com/Cerebellum-Network/cere-media-sdk-js/compare/v1.4.2-rc.13...v1.4.2-rc.14) (2024-08-28)
 
-**Note:** Version bump only for package @cere/media-sdk-client
-
 ## [1.4.2-rc.13](https://github.com/Cerebellum-Network/cere-media-sdk-js/compare/v1.4.2-rc.12...v1.4.2-rc.13) (2024-08-27)
 
+## [1.4.2-rc.12](https://github.com/Cerebellum-Network/cere-media-sdk-js/compare/v1.4.2-rc.11...v1.4.2-rc.12) (2024-08-27)
+
 **Note:** Version bump only for package @cere/media-sdk-client
 
-## [1.4.2-rc.12](https://github.com/Cerebellum-Network/cere-media-sdk-js/compare/v1.4.2-rc.11...v1.4.2-rc.12) (2024-08-27)
+## [1.5.2-rc.0](https://github.com/Cerebellum-Network/cere-media-sdk-js/compare/v1.4.2-rc.11...v1.5.2-rc.0) (2024-11-14)
+
+**Note:** Version bump only for package @cere/media-sdk-client
+
+## [1.4.2-rc.15](https://github.com/Cerebellum-Network/cere-media-sdk-js/compare/v1.4.2-rc.11...v1.4.2-rc.15) (2024-11-14)
+
+**Note:** Version bump only for package @cere/media-sdk-client
+
+## [1.4.2-rc.14](https://github.com/Cerebellum-Network/cere-media-sdk-js/compare/v1.4.2-rc.11...v1.4.2-rc.14) (2024-11-14)
+
+**Note:** Version bump only for package @cere/media-sdk-client
+
+## [1.4.2-rc.13](https://github.com/Cerebellum-Network/cere-media-sdk-js/compare/v1.4.2-rc.11...v1.4.2-rc.13) (2024-11-14)
+
+**Note:** Version bump only for package @cere/media-sdk-client
+
+## [1.4.2-rc.12](https://github.com/Cerebellum-Network/cere-media-sdk-js/compare/v1.4.2-rc.11...v1.4.2-rc.12) (2024-11-14)
 
 **Note:** Version bump only for package @cere/media-sdk-client